pyproject.toml
[tool.poetry]
name = "djangorabbit"
version = "0.1.0"
description = ""
authors = ["edummap4412 <mascarenhaseduardo4412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
celery = "^5.3.1"
django = "^4.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
